chore: purpose improve

+++ packages/plugin-loans-api/src/models/loanPurpose.ts
@@ -6,6 +6,7 @@ import {
 import { Model } from 'mongoose';
 import { IModels } from '../connectionResolver';
 import { FilterQuery } from 'mongoose';
+import { escapeRegExp } from '@erxes/api-utils/src/core';
 
 export interface IPurposeModel extends Model<IPurposeDocument> {
   getPurpose(selector: FilterQuery<IPurposeDocument>);
@@ -34,14 +35,49 @@ export const loadPurposeClass = (models: IModels) => {
      * Create a Purpose
      */
     public static async createPurpose(doc: IPurpose) {
-      return models.LoanPurpose.create(doc);
+      const parentCategory = await models.LoanPurpose.findOne({
+        _id: doc.parentId,
+      }).lean();
+
+      doc.order = await this.generateOrder(parentCategory, doc);
+
+      return models.LoanPurpose.create({ ...doc, createdAt: new Date() });
     }
 
     /**
      * Update Purpose
      */
     public static async updatePurpose(_id: string, doc: IPurpose) {
+      const category = await models.LoanPurpose.getPurpose({
+        _id,
+      });
+
+      const parentCategory = await models.LoanPurpose.findOne({
+        _id: doc.parentId,
+      }).lean();
+
+      doc.order = await this.generateOrder(parentCategory, doc);
+
+      const childCategories = await models.LoanPurpose.find({
+        $and: [
+          { order: { $regex: new RegExp(`^${escapeRegExp(category.order)}`) } },
+          { _id: { $ne: _id } },
+        ],
+      });
+
       await models.LoanPurpose.updateOne({ _id }, { $set: doc });
+
+      // updating child categories order
+      childCategories.forEach(async (childCategory) => {
+        let order = childCategory.order;
+
+        order = order.replace(category.order, doc.order);
+
+        await models.LoanPurpose.updateOne(
+          { _id: childCategory._id },
+          { $set: { order } }
+        );
+      });
 
       return models.LoanPurpose.findOne({ _id });
     }
@@ -50,7 +86,41 @@ export const loadPurposeClass = (models: IModels) => {
      * Remove Purpose
      */
     public static async removePurposes(_ids: string[]) {
+      // Step 1: Check if all LoanPurpose entries exist
+      const existingPurposes = await models.LoanPurpose.find({
+        _id: { $in: _ids },
+      });
+      if (existingPurposes.length !== _ids.length) {
+        throw new Error('Some Loan Purposes do not exist');
+      }
+
+      // Step 2: Check if any active Loans reference these purposes
+      const count = await models.LoanPurpose.countDocuments({
+        parentId: { $in: _ids },
+      });
+
+      if (count > 0) {
+        throw new Error(
+          "Can't remove purposes that are being used in child purpose"
+        );
+      }
+
+      // Step 3: Delete the purposes
       return models.LoanPurpose.deleteMany({ _id: { $in: _ids } });
+    }
+
+    /**
+     * Generating order
+     */
+    public static async generateOrder(
+      parentCategory: IPurpose | null | undefined,
+      doc: IPurpose
+    ) {
+      const order = parentCategory
+        ? `${parentCategory.order}${doc.code}/`
+        : `${doc.code}/`;
+
+      return order;
     }
   }
